Make local-setup depend on yq and dev-tls

+++ Makefile
@@ -63,7 +63,7 @@ test: manifests generate fmt vet envtest ## Run tests.
 	KUBEBUILDER_ASSETS="$(shell $(ENVTEST) use $(ENVTEST_K8S_VERSION) --bin-dir $(LOCALBIN) -p path)" go test ./... -coverprofile cover.out
 
 .PHONY: local-setup
-local-setup: kind kustomize helm ## Setup multi cluster traffic controller locally using kind.
+local-setup: kind kustomize helm yq dev-tls ## Setup multi cluster traffic controller locally using kind.
 	./hack/local-setup.sh
 
 ##@ Build
@@ -133,7 +133,7 @@ deploy-sample-applicationset:
 
 .PHONY: dev-tls
 dev-tls:
-	openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout config/webhook-setup/tls/tls.key -out config/webhook-setup/tls/tls.crt -subj "/C=IE/O=Red Hat Ltd/OU=HCG/CN=webhook.172.18.0.2.nip.io"
+	test -s config/webhook-setup/control/tls/tls.crt || openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout config/webhook-setup/control/tls/tls.key -out config/webhook-setup/control/tls/tls.crt -subj "/C=IE/O=Red Hat Ltd/OU=HCG/CN=webhook.172.18.0.2.nip.io"
 
 ##@ Build Dependencies
 
@@ -190,5 +190,5 @@ $(HELM):
 
 .PHONY: yq
 yq: $(YQ)
-	GOBIN=$(LOCALBIN) go install github.com/mikefarah/yq/v4@$(YQ_VERSION)
+	test -s $(LOCALBIN)/yq || GOBIN=$(LOCALBIN) go install github.com/mikefarah/yq/v4@$(YQ_VERSION)
